+++ docs/changelog.rst
@@ -1,13 +1,16 @@
 ChangeLog
 =========
 
+.. currentmodule:: arlib
+
 0.0.4
 -----
 * Add :func:`arlib.open` as a shortcut of :class:`Archive` constructor
   (:issue:`1`, :pr:`2`).
-* Add :func:`arlib.is_archive` to determine if a file could be opened
+* Add :func:`is_archive` to determine if a file could be opened
   as a valid archive (:issue:`3`, :pr:`4`).
-  
+* Add :func:`assert_is_archive` (:pr:`5`).
+
 0.0.3
 -----